Remove BitUtilBig (#2811)
* Fix interchanged parameter names in bit util

* removed BitUtilBig

* cleanup

---------
