+++ core/src/main/java/com/graphhopper/routing/ev/TurnCost.java
@@ -1,7 +1,7 @@
 package com.graphhopper.routing.ev;
 
 import com.graphhopper.storage.IntsRef;
-import com.graphhopper.util.Helper;
+import com.graphhopper.util.BitUtil;
 
 import static com.graphhopper.routing.util.EncodingManager.getKey;
 
@@ -15,7 +15,7 @@ public class TurnCost {
      * This creates an EncodedValue specifically for the turn costs
      */
     public static DecimalEncodedValue create(String name, int maxTurnCosts) {
-        int turnBits = Helper.countBitValue(maxTurnCosts);
+        int turnBits = BitUtil.countBitValue(maxTurnCosts);
         return new DecimalEncodedValueImpl(key(name), turnBits, 0, 1, false, false, true);
     }
