+++ web-api/src/test/java/com/graphhopper/util/HelperTest.java
@@ -32,17 +32,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 public class HelperTest {
 
     @Test
-    public void testCountBitValue() {
-        assertEquals(1, Helper.countBitValue(1));
-        assertEquals(2, Helper.countBitValue(2));
-        assertEquals(2, Helper.countBitValue(3));
-        assertEquals(3, Helper.countBitValue(4));
-        assertEquals(3, Helper.countBitValue(7));
-        assertEquals(4, Helper.countBitValue(8));
-        assertEquals(5, Helper.countBitValue(20));
-    }
-
-    @Test
     public void testGetLocale() {
         assertEquals(Locale.GERMAN, Helper.getLocale("de"));
         assertEquals(Locale.GERMANY, Helper.getLocale("de_DE"));
@@ -67,28 +56,6 @@ public class HelperTest {
         assertEquals(2, Helper.keepIn(2, 1, 4), 1e-2);
         assertEquals(3, Helper.keepIn(2, 3, 4), 1e-2);
         assertEquals(3, Helper.keepIn(-2, 3, 4), 1e-2);
-    }
-
-    @Test
-    public void testUnsignedConversions() {
-        long l = Helper.toUnsignedLong(-1);
-        assertEquals(4294967295L, l);
-        assertEquals(-1, Helper.toSignedInt(l));
-
-        int intVal = Integer.MAX_VALUE;
-        long maxInt = (long) intVal;
-        assertEquals(intVal, Helper.toSignedInt(maxInt));
-
-        intVal++;
-        maxInt = Helper.toUnsignedLong(intVal);
-        assertEquals(intVal, Helper.toSignedInt(maxInt));
-
-        intVal++;
-        maxInt = Helper.toUnsignedLong(intVal);
-        assertEquals(intVal, Helper.toSignedInt(maxInt));
-
-        assertEquals(0xFFFFffffL, (1L << 32) - 1);
-        assertTrue(0xFFFFffffL > 0L);
     }
 
     @Test
